refactor backend code to routes, controllers, and repositories specific for user
