+++ client/src/components/welcome/Welcome.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Welcome = () => {
+  return;
+  <div className="welcome-page">
+    <h1 className="title">Welcome to your Password Manager</h1>
+    <h2 className="instructions">
+      Please, create a new account, or log in with with your credentials{" "}
+    </h2>
+  </div>;
+};
+
+export default Welcome;
