+++ client/src/styles/welcome.css
@@ -1,0 +1,10 @@
+.welcome-page {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  text-align: center;
+}
+
+.title {
+  font-size: 2rem;
+}
